backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
django-cors-headers
googlemaps
celery
django-celery-beat
redis
